refactor: change CollisionSystem to use marker logic
