+++ core/src/main/kotlin/com/tdt4240/group3/controller/systems/TroopCreationSystem.kt
@@ -13,13 +13,15 @@ import com.tdt4240.group3.model.components.marker.SelectableComponent
 import com.tdt4240.group3.model.entities.EntityFactory
 import ktx.ashley.allOf
 import ktx.ashley.get
+import kotlin.compareTo
+import kotlin.text.get
 
 class TroopCreationSystem(private val engine: Engine) : EntitySystem() {
-
     private val factory = EntityFactory(engine)
     private val cityFamily = allOf(CityComponent::class, PositionComponent::class, TeamComponent::class).get()
     private val gameStateFamily = allOf(GameStateComponent::class, NeedsTroopSpawnComponent::class).get()
     private val troopFamily = allOf(TroopComponent::class, TeamComponent::class).get()
+
 
     override fun update(deltaTime: Float) {
         val gameStateEntity = engine.getEntitiesFor(gameStateFamily).firstOrNull() ?: return
@@ -30,16 +32,36 @@ class TroopCreationSystem(private val engine: Engine) : EntitySystem() {
         gameStateEntity.remove(NeedsTroopSpawnComponent::class.java)
     }
 
-    fun createTroopFromCity(cityEntity: Entity): Entity {
-        return factory.createTroopFromCity(cityEntity)
+    fun createTroopFromCity(cityEntity: Entity) {
+        val cityComp = cityEntity[CityComponent.mapper] ?: return
+        val cityPos = cityEntity[PositionComponent.mapper] ?: return
+        val cityTeam = cityEntity[TeamComponent.mapper]?.team ?: return
+        val existingTroop = engine.getEntitiesFor(troopFamily).find {
+            val p = it[PositionComponent.mapper]
+            p?.q == cityPos.q && p.r == cityPos.r
+        }
+        if (existingTroop != null) {
+            val troopComp = existingTroop[TroopComponent.mapper]!!
+            val troopTeam = existingTroop[TeamComponent.mapper]?.team ?: return
+            if (troopTeam == cityTeam) {
+                val total = troopComp.strength + cityComp.baseProduction
+                if (total <= 99) {
+                    troopComp.strength = total
+                }
+                else {
+                    troopComp.strength = 99
+                }
+            }
+        } else {
+            val newTroop = factory.createTroopFromCity(cityEntity)
+            newTroop.add(engine.createComponent(SelectableComponent::class.java))
+        }
     }
 
     fun createTroopsForTeam(team: TeamComponent.TeamName) {
-    engine.getEntitiesFor(cityFamily)
-        .filter { TeamComponent.mapper.get(it)?.team == team }
-        .forEach {
-            createTroopFromCity(it)
-        }
+        engine.getEntitiesFor(cityFamily)
+            .filter { TeamComponent.mapper.get(it)?.team == team }
+            .forEach { createTroopFromCity(it) }
     }
 
     fun markSelectable(gs: GameStateComponent) {
