+++ core/src/main/kotlin/com/tdt4240/group3/model/components/TroopComponent.kt
@@ -7,18 +7,8 @@ import ktx.ashley.mapperFor
 
 class TroopComponent : Component, Pool.Poolable {
     var strength: Int = 0
-    var isColliding: Boolean = false
-
-    fun colliding(){
-        this.isColliding = true
-    }
 
     override fun reset() {
-        isColliding = false
-    }
-
-    fun resetForNewTurn(){
-        isColliding = false
     }
 
     companion object {
