+++ core/src/main/kotlin/com/tdt4240/group3/model/components/marker/CollidingComponent.kt
@@ -1,0 +1,8 @@
+package com.tdt4240.group3.model.components.marker
+
+import com.badlogic.ashley.core.Component
+import com.badlogic.gdx.utils.Pool
+
+class CollidingComponent: Component, Pool.Poolable {
+    override fun reset() {}
+}
